// src/api/model/userPreferences.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReceiptStatus } from './receiptStatus';


export interface UserPreferences { 
    /**
     * User preferences id
     */
    id: number;
    createdAt: string;
    createdBy?: number;
    /**
     * Created by entity\'s name
     */
    createdByString?: string;
    updatedAt?: string;
    /**
     * User foreign key
     */
    userId: number;
    /**
     * Group foreign key
     */
    quickScanDefaultGroupId?: number;
    /**
     * User foreign key
     */
    quickScanDefaultPaidById?: number;
    quickScanDefaultStatus?: ReceiptStatus;
}